[loss]: compile loss
