+++ touchnet/models/llama/__init__.py
@@ -1,61 +1,19 @@
 from typing import Tuple
 
 import torch
-from torch.distributed.tensor import DTensor, Replicate
 from transformers.models.llama import LlamaConfig, LlamaForCausalLM
 from transformers.models.llama.modeling_llama import LlamaRMSNorm
 
 from touchnet.data.dataloader import build_dataloader
+from touchnet.loss.cross_entropy import cross_entropy_loss
 from touchnet.models.llama.configuration_llama import LlamaForASRConfig
 from touchnet.models.llama.modeling_llama import LlamaForASR
 from touchnet.models.llama.parallelize_llama import parallelize_llama
 from touchnet.models.llama.pipeline_llama import pipeline_llama
 from touchnet.tokenizer.tokenizer import build_tokenizer
-from touchnet.utils.metrics import build_metrics_processor
+from touchnet.utils.metrics import accuracy, build_metrics_processor
 from touchnet.utils.optimizer import build_lr_schedulers, build_optimizers
 from touchnet.utils.train_spec import TrainSpec, register_train_spec
-
-
-def cross_entropy_loss(
-    pred: torch.Tensor, labels: torch.Tensor,
-    sentence_lens: torch.Tensor, num_sentence: int,
-    ignore_index: int = -100,
-) -> Tuple[torch.Tensor, torch.Tensor]:
-    """Common cross-entropy loss function for Transformer models training.
-
-    Args:
-        pred (Tensor): Prediction tensors (B, Lmax // cp, Vocab // tp) if pred.to_local()
-                       else (B, Lmax // cp, Vocab) if pred.full_tensor()
-        labels (LongTensor): Target label tensors (B, Lmax // cp).
-        ignore_index (int): Ignore label id.
-
-    Returns:
-        Tuple[torch.Tensor, torch.Tensor]: Loss value.
-    """
-
-    if isinstance(pred, DTensor):
-        # NOTE(xcsong): make sentence_lens distributed to work with DTensor-style Loss
-        sentence_lens = DTensor.from_local(
-            sentence_lens, pred.device_mesh, [Replicate()], run_check=False
-        )  # (bs, seq_len // cp)
-    batch_size = pred.size(0)
-    num_tokens = (labels != ignore_index).sum().item()
-    loss = torch.nn.functional.cross_entropy(
-        pred.flatten(0, 1).float(), labels.flatten(0, 1),
-        reduction="none", ignore_index=ignore_index,
-    )  # (bs * seq_len // cp,)
-    # NOTE(xcsong): per-sample loss for backward while per-token loss for logging.
-    loss_per_token = loss.sum()
-    if loss_per_token > 1e-6 and num_tokens > 0:
-        loss_per_token = (loss.sum() / num_tokens)  # (1,)
-    else:
-        loss_per_token = torch.zeros_like(loss_per_token)
-    loss_per_sample = loss.reshape(batch_size, -1)  # (bs, seq_len // cp)
-    # 1. reduce loss over sentence
-    loss_per_sample = torch.sum(loss_per_sample / sentence_lens, dim=-1)  # (bs,)
-    # 2. reduce loss over global-batch
-    loss_per_sample = torch.sum(loss_per_sample) / num_sentence  # (1,)
-    return loss_per_sample, loss_per_token
 
 
 def post_init(model: LlamaForCausalLM, init_device: torch.device):
@@ -79,33 +37,6 @@ def post_init(model: LlamaForCausalLM, init_device: torch.device):
     for name, param in model.named_parameters():
         if torch.isnan(param).any() or torch.isinf(param).any():
             raise ValueError(f"NaN/inf in model parameters `{name}`.")
-
-
-def accuracy(pred: torch.Tensor, labels: torch.Tensor,
-             ignore_index: int = -100) -> torch.Tensor:
-    """Calculate accuracy.
-
-    Args:
-        pred (Tensor): Prediction tensors (B, Lmax // cp, Vocab // tp) if pred.to_local()
-                       else (B, Lmax // cp, Vocab) if pred.full_tensor()
-        labels (LongTensor): Target label tensors (B, Lmax // cp).
-        ignore_index (int): Ignore label id.
-
-    Returns:
-        torch.Tensor: Accuracy value (0.0 - 1.0).
-
-    """
-    if isinstance(pred, DTensor):
-        pred = pred.full_tensor()  # (B, T // cp, V)
-    pred = pred.argmax(dim=-1)  # (B, T // cp, V) -> (B, T // cp)
-    mask = labels != ignore_index
-    numerator = torch.sum(
-        pred.masked_select(mask) == labels.masked_select(mask))
-    denominator = torch.sum(mask)
-    if denominator > 0:
-        return (numerator / denominator).detach()
-    else:
-        return torch.zeros_like(numerator).detach()
 
 
 def get_num_flop_per_token(num_params: int, model_config: LlamaConfig, seq_len: int) -> int:
